modified:   index.php 	modified:   pengeluaran.php

+++ pengeluaran.php
@@ -135,7 +135,9 @@
                             </form>
                           </div>
                         </div>
-                        <a href="hapus_pengeluaran.php?id_pengeluaran=<?= $dp['id_pengeluaran']; ?>" class="badge badge-danger btn-delete" data-nama="Pengeluaran : Rp. <?= number_format($dp['jumlah_pengeluaran']); ?> | <?= $dp['keterangan']; ?>"><i class="fas fa-fw fa-trash"></i> Hapus</a>
+                        <?php if ($_SESSION['id_jabatan'] == '1'): ?>
+                          <a href="hapus_pengeluaran.php?id_pengeluaran=<?= $dp['id_pengeluaran']; ?>" class="badge badge-danger btn-delete" data-nama="Pengeluaran : Rp. <?= number_format($dp['jumlah_pengeluaran']); ?> | <?= $dp['keterangan']; ?>"><i class="fas fa-fw fa-trash"></i> Hapus</a>
+                        <?php endif ?>
                       </td>
                     </tr>
                   <?php endforeach ?>
